feat : Mobile 및 Desktop 헤더 분리
